services/analytics-service/requirements.txt
requests
paho-mqtt
PyYAML
Flask 